+++ CHANGELOG.md
@@ -19,6 +19,7 @@
  - Fix #242: Add POST filter for `admin/force-password-change` action (bscheshirwork)
  - Fix #252: Delete check for unexpected property `allowPasswordRecovery` for resend email by admin (bscheshirwork)
  - Fix #254: Rename `GDPR` properties to `lowerCamelCase` style (bscheshirwork)
+ - Fix #271: Add closure support for `from` email address; Change default sender to `supportEmail` (bscheshirwork)
 
 ## 1.1.4 - February 19, 2018
 - Enh: Check enableEmailConfirmation on registration (faenir)